CMSDK : Fix generateAlertColorFromDrawable for non BitmapDrawable
In the case that the bitmap being passed in was not a bitmap drawable,
we were not retaining any of the attributes from the original drawable.
This patch ensures we ask that drawable to draw on the canvas/bitmap so we
can use that information. Also add tests around it.

Change-Id: I3eefba6e6624fe0bed4965ddf9029320c40f7420
Reviewed-on: http://review.blissroms.com/3869
Reviewed-by: Vaughn N <rwaterspf1@floydco.net>
Reviewed-by: blueyes <brian.w.petroski@gmail.com>
Tested-by: blueyes <brian.w.petroski@gmail.com>